Document generation process organized into classes
